fix(quantity): properly convert cores
Sometimes the value is stored as cores, and sometimes it is stored as
millicores. So just calling .Value() doesn't seem to give consistent
results. This checks the suffix to ensure that we convert cores the
correct way.

Its possible that we may need to do this on Memory as well, but so far I
haven't seen any negative impacts.